+++ src/fr/maxlego08/menu/loader/RequirementLoader.java
@@ -15,7 +15,6 @@ import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 public class RequirementLoader implements Loader<Requirement> {
 
@@ -31,10 +30,9 @@ public class RequirementLoader implements Loader<Requirement> {
         File file = (File) objects[0];
         ButtonManager buttonManager = this.plugin.getButtonManager();
         List<Permissible> permissibles = buttonManager.loadPermissible((List<Map<String, Object>>) configuration.getList(path + "requirements", configuration.getList(path + "requirement", new ArrayList<>())), path, file);
-        List<Action> successActions = buttonManager.loadActions((List<Map<String, Object>>) configuration.getList(path + "success", new ArrayList<>()), path+"success", file);
-        List<Action> denyActions = buttonManager.loadActions((List<Map<String, Object>>) configuration.getList(path + "deny", new ArrayList<>()), path+"deny", file);
-        List<ClickType> clickTypes = configuration.getStringList(path + "clicks").stream().map(String::toUpperCase)
-                .map(ClickType::valueOf).collect(Collectors.toList());
+        List<Action> successActions = buttonManager.loadActions((List<Map<String, Object>>) configuration.getList(path + "success", new ArrayList<>()), path + "success", file);
+        List<Action> denyActions = buttonManager.loadActions((List<Map<String, Object>>) configuration.getList(path + "deny", new ArrayList<>()), path + "deny", file);
+        List<ClickType> clickTypes = this.plugin.getInventoryManager().loadClicks(configuration.getStringList(path + "clicks"));
         int miniumRequirement = configuration.getInt(path + "miniumRequirement", permissibles.size());
 
         return new ZRequirement(miniumRequirement, permissibles, denyActions, successActions, clickTypes);