+++ changelog.md
@@ -26,6 +26,15 @@
 
 # Unreleased
 
+# 1.0.2.0
+
+- Added ANY and ALL support for click requirements https://github.com/Maxlego08/zMenu/issues/24 
+- Added a check when displaying the item to prevent that if the slot is outside the slots this triggers an error but just a message in the console
+- Fix error with 1.20.4
+- Fix error with Meta
+- Fix error with slot page and else button 
+- Fix checkPermission with view requirement who made it check placeholders and permissions impossible
+
 # 1.0.1.9
 
 - Added back type for action requirements